requirement.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
seaborn
seaborn
pillow
nltk
python-Levenshtein
collections
PIL
os
math
re
pickle
random
torch.nn